requirements.txt
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
protobuf==3.20.3
Requests==2.31.0
streamlit==1.22.0
torch==2.0.1
transformers==4.29.2
